+++ pyoffers/models/goal.py
@@ -23,5 +23,7 @@ class GoalManager(ModelManager):
     def find_by_id(self, id):
         return self._call('findById', id=id)
 
-    def find_all(self, sort=None, limit=None, page=None, **kwargs):
-        return self._call('findAll', filters=kwargs, sort=sort, limit=limit, page=page, single_result=False)
+    def find_all(self, sort=None, limit=None, page=None, contain=None, **kwargs):
+        return self._call(
+            'findAll', filters=kwargs, sort=sort, limit=limit, page=page, contain=contain, single_result=False
+        )